refactor: データモデル構造体を`models.rs`に集約し、`main.rs`、`reports.rs`、`stats.rs`から参照するように変更
